+++ apps/api/src/pagination/S3ListSource.ts
@@ -1,7 +1,5 @@
-import { ListPageVariant } from "../entities/getListJSONKey"
-
 export type S3ListSource = Readonly<{
     getIndexKey: () => string
-    getPageKey: (pageIndex: number, variant: ListPageVariant) => string
+    getPageKey: (pageIndex: number) => string
     isEligible: (listQuery: Readonly<Record<string, string | number | boolean | undefined>>) => boolean
 }>